% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_dsatur}
\alias{color_dsatur}
\alias{color_msc}
\alias{color_lmxrlf}
\alias{color_hybrid_tabucol}
\alias{color_hybrid_dsatur}
\title{Color Graph}
\usage{
color_dsatur()

color_msc()

color_lmxrlf()

color_hybrid_tabucol()

color_hybrid_dsatur()
}
\description{
Color Graph
}
\examples{
library(tidygraph)
library(igraph)

play_islands(5, 10, 0.8, 3) \%>\%
  mutate(color = color_dsatur())

}
